// pagonxt-onetradefinance-external-frontend/src/api/types/cli/otherOperations/CreateFormDataDto.ts
import { CustomerCliRequestDto } from '../../CustomerCliRequestDto';
import { CustomerDto } from '../../CustomerDto';
import { FormDocumentationDto } from '../../forms/FormDocumentationDto';

export interface CreateFormDataDto {
  code?: string | null;
  customer: CustomerDto;
  documentation: FormDocumentationDto;
  request: {
    comments: string;
    importCollection: CustomerCliRequestDto;
    office: string;
    operationType: string;
  };
  savedStep: number;
}